requirements.txt
Flask-RESTful==0.3.6
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.3.2
PyMySQL==0.9.3
Werkzeug==0.14.1
# pip install -r requirements.txt
